# api/cmd/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "properties": {
          "error": {
            "description": "Human-readable error message",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "pds_... API token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "url": "https://pukucloud.ai"
    },
    "description": "REST API for PukuCloud — Firecracker microVM sandboxes and managed Postgres. Authenticate with a pds_ bearer token from the dashboard's API Tokens page.",
    "title": "PukuCloud API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/sandboxes/{id}/fs": {
      "delete": {
        "operationId": "deletePath",
        "parameters": [
          {
            "description": "Sandbox UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Absolute guest path to delete",
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sandbox not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Delete failed"
          }
        },
        "summary": "Delete a file or directory in the guest",
        "tags": [
          "Sandboxes"
        ]
      },
      "get": {
        "operationId": "readFile",
        "parameters": [
          {
            "description": "Sandbox UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Absolute guest path to read",
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "File contents"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sandbox not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Read failed (e.g. no such file)"
          }
        },
        "summary": "Read a file's raw bytes from the guest filesystem",
        "tags": [
          "Sandboxes"
        ]
      },
      "put": {
        "operationId": "writeFile",
        "parameters": [
          {
            "description": "Sandbox UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Absolute guest path to write",
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "description": "Raw file bytes, max 32 MiB",
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "bytes": {
                      "description": "Byte count written",
                      "type": "integer"
                    },
                    "path": {
                      "description": "Guest path written",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Written"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Body too large (>32 MiB) or unreadable"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sandbox not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Write failed"
          }
        },
        "summary": "Write raw request-body bytes to a guest file (32 MiB cap)",
        "tags": [
          "Sandboxes"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://api.pukucloud.ai"
    }
  ]
}